+++ extensions/copyq-clipboard-manager/CHANGELOG.md
@@ -1,9 +1,11 @@
 # CopyQ Changelog
 
-## [Version 2.0] - 2023-11-18
+## [Version 2.0] - 2023-11-28
 
 - Add second command to choose between a list of CopyQ tabs
+- Changed clipboard contents display to show shortened and full content simultaneously
 - Performance upgrades, allowing the user to display more clipboard contents faster.
+- Fixed error handling of CopyQ not found and CopyQ server not running.
 
 ## [Added CopyQ Clipboard Manager] - 2023-11-01